Moves newGameIntro to characterCreation folder
